requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
llama-index-readers-file
chromadb
sqlalchemy
python-multipart
python-dotenv
pytest
httpx
python-jose[cryptography]
passlib[argon2]
